1_Flask_Backend/flask-backend/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
blurhash==1.1.4
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
CouchDB==1.2
decorator==5.1.1
Flask==2.0.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
Mastodon.py==1.8.1
nltk==3.8.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
regex==2023.5.5
requests==2.29.0
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.0.3
